logs e ponto de equilibrio

+++ src/main/java/senai/CursosFic/model/Curso.java
@@ -33,7 +33,7 @@ public class Curso {
 	private TipoAtendimento tipoAtendimento;
 	@Enumerated(EnumType.STRING)
 	private Nivel nivel;
-	private int cargaHoraria;
+	private Double cargaHoraria;
 	@ManyToOne
 	private Area area;
 	private Double valor;
